+++ commune_app/models.py
@@ -1,3 +1,3 @@
 from commune_app.all_models.users import User  # noqa:F401
-from commune_app.all_models.Chore import Chore  # noqa:F401
-from commune_app.all_models.Vote import Vote  # noqa:F401
+from commune_app.all_models.chores import Chore  # noqa:F401
+from commune_app.all_models.votes import Vote  # noqa:F401
